requirements.txt
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
